Update stories to use flex-wrap
